+++ lib/scoring/index.ts
@@ -27,6 +27,7 @@ import * as ambrose from './modes/ambrose';
 import * as floridaScramble from './modes/floridaScramble';
 import * as fourballMatchplay from './modes/fourballMatchplay';
 import * as foursomesMatchplay from './modes/foursomesMatchplay';
+import * as greensomeMatchplay from './modes/greensomeMatchplay';
 import * as wolf from './modes/wolf';
 import * as nassau from './modes/nassau';
 import * as skins from './modes/skins';
@@ -60,6 +61,8 @@ export function computeLeaderboard(ctx: ScoringContext): ModeResult {
       return fourballMatchplay.compute(ctx);
     case 'foursomes_matchplay':
       return foursomesMatchplay.compute(ctx);
+    case 'greensome_matchplay':
+      return greensomeMatchplay.compute(ctx);
     case 'wolf':
       return wolf.compute(ctx);
     case 'nassau':